feat(apps/kampus): move @kampus/relay into ~/features/relay
