+++ apps/kampus/app/sozluk/[term]/SozlukTerm.tsx
@@ -2,8 +2,8 @@
 
 import { graphql, usePreloadedQuery } from "react-relay";
 
-import { type SerializablePreloadedQuery } from "@kampus/relay";
-import useSerializablePreloadedQuery from "@kampus/relay/use-serializable-preloaded-query";
+import { type SerializablePreloadedQuery } from "features/relayfeatures/relay";
+import useSerializablePreloadedQuery from "features/relayfeatures/relayfeatures/relay/use-serializable-preloaded-query";
 
 import { type SozlukTermQuery } from "./__generated__/SozlukTermQuery.graphql";
 import { SozlukTermBody } from "./SozlukTermBody";
