+++ apps/kampus/app/sozluk/[term]/page.tsx
@@ -1,5 +1,4 @@
-import loadSerializableQuery from "@kampus/relay/load-serializable-query";
-
+import loadSerializableQuery from "~/features/relay/load-serializable-query";
 import { SozlukTermContainer } from "~/app/sozluk/[term]/SozlukTerm";
 import SozlukTermQueryNode, { type SozlukTermQuery } from "./__generated__/SozlukTermQuery.graphql";
