+++ apps/kampus/features/relay/RelayEnvironmentProvider.tsx
@@ -3,7 +3,7 @@
 import { type PropsWithChildren } from "react";
 import { RelayEnvironmentProvider as ReactRelayEnvironmentProvider } from "react-relay";
 
-import { getCurrentEnvironment } from "@kampus/relay";
+import { getCurrentEnvironment } from "./environment";
 
 export const RelayEnvironmentProvider = ({ children }: PropsWithChildren) => {
   const environment = getCurrentEnvironment();
