+++ apps/kampus/features/relay/environment.ts
@@ -1,4 +1,4 @@
-import { type ReadonlyRequestCookies } from "next/dist/server/web/spec-extension/adapters/request-cookies";
+import { auth } from "@clerk/nextjs";
 import {
   Environment,
   Network,
@@ -32,9 +32,16 @@ export async function networkFetch(
   // but on next.js servers, `fetch` is pollyfilled, and does not pick up cookies
   // we use `next/header`'s cookies() helper here
   if (IS_SERVER) {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const nextHeaders = require("next/headers") as { cookies: () => ReadonlyRequestCookies };
-    headers.Cookie = nextHeaders.cookies().toString();
+    const { getToken } = auth();
+
+    // console.log(">>>>>>>>>>>>>>>>>>>", nextHeaders.headers());
+    // console.log("request headers >>>>>>>>", request.headers);
+
+    const token = await getToken();
+
+    console.log("naber", { token });
+
+    headers.Authorization = `Bearer ${token}`;
   }
 
   const resp = await fetch(HTTP_ENDPOINT, {
@@ -93,6 +100,7 @@ function createNetwork() {
 }
 
 function createEnvironment() {
+  console.log("creating environment", { IS_SERVER });
   return new Environment({
     network: createNetwork(),
     store: new Store(RecordSource.create()),
