+++ apps/kampus/features/relay/load-serializable-query.ts
