+++ apps/kampus/features/relay/use-serializable-preloaded-query.ts
