+++ apps/kampus/package.json
@@ -14,17 +14,20 @@
   },
   "devDependencies": {
     "@kampus/tailwind": "*",
-    "@types/react-relay": "14.1.4"
+    "@types/react-relay": "14.1.4",
+    "@types/relay-runtime": "14.1.10"
   },
   "dependencies": {
     "@clerk/localizations": "1.26.11",
     "@clerk/nextjs": "4.27.2",
     "@clerk/themes": "1.7.9",
-    "@kampus/relay": "*",
+    "@kampus/std": "*",
     "@kampus/ui": "*",
-    "react-markdown": "^9.0.1",
-    "react-relay": "15.0.0",
+    "@radix-ui/react-icons": "1.3.0",
+    "@radix-ui/themes": "2.0.3",
     "relay-compiler": "15.0.0",
+    "react-relay": "15.0.0",
+    "relay-runtime": "15.0.0",
     "znv": "0.3.2",
     "zod": "3.21.4"
   }
