--- packages/relay/.gitignore
@@ -1,1 +1,0 @@
-.contentlayer
